requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
google-generativeai==0.3.2
boto3==1.34.10
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
ecdsa==0.18.0
python-multipart==0.0.6